Add Dependencies.xcodeproj as subproject of app xcodeproj
Currently this has to be done using Xcode 8, because adding ANY kind of
subproject within Xcode 9 causes a SIGILL crash inside
IDEFoundation.IDEStructureEditingProcedure.performTimeslicedSyncOnMainThread<A>()

Fortunately, the format is the same across both versions, and Xcode 9 is
perfectly happy to use a project with subprojects - just not add them.

I haven't linked to any dependencies yet, because at the moment we only
have our empty Dependencies.framework - we will do that step once some
dependencies are added at the SPM level.

This does mean that the app xcodeproj will be broken if you open it
without first running `generate-dependencies-project.rb`